+++ changes.md
@@ -1,5 +1,8 @@
 ## v0.1.x - 2016-02-160
 - Add support of read/write netCDF point data following CF conventions
+- Add support for disabling automatic masking during reading. Useful if the data
+  has fill values but needs to be scaled to a datatype that does not support NaN
+  values.
 
 ## v0.1.6 - 2016-02-03
 - Add support for disabling automatic scaling in base netCDF4 library.